add confident score

+++ train_val.py
@@ -75,7 +75,7 @@ elif torch.backends.mps.is_available():
 else:
     print('Using CPU for training')
     device = torch.device('cpu')
-dtype = torch.float16
+dtype = torch.float32
 intention_class = ['interacting', 'interested', 'not_interested']
 attitude_classes = ['positive', 'negative', 'no_interacting']
 action_classes = ['hand_shake', 'hug', 'pet', 'wave', 'punch', 'throw', 'point-converse', 'gaze', 'leave',
